requirements.txt
asgiref==3.8.1
contourpy==1.3.0
crispy-bootstrap4==2024.1
cycler==0.12.1
Django==5.1.1
django-crispy-forms==2.3
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
mpmath==1.3.0
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
